[test] Remove the check on default charset as we are not testing what java does
this breaks probably due to something being set via maven but is fine in IDE after changes.  Since this is read from system properly but this setting is clear default now, its clear something in framework is changing values on the fly that would not have been affected before but also was not needed.