[rework] Use 'row' throughout instead of object and fix test sample to extend row
This shows that one can simply use row as base class and use that instead which I think fixes what some users were asking for.  The internals were all 'object' when I got it which is sloppy at best.  I'm not sure why we would have anything but some form of row since everything denotes that.