Cleranup coding conventions doc
code is auto formatted so most of this doesn't matter.  Removed legacy author/version as version doesn't apply to git and author incorrectly applies author is solely responsible for the class which is never true over time and git has this data in far more accurate way.  We never want to blame an author which will happen regardless of touching if they are only one listed, plus its a lot of overhead.