Force paginated lists to define the type as <Row>
presumably the ask from users was to not use the Row object which they can do so in their own definitions now that use paginated list.  Its going to be Row otherwise as it stands or it has type safety issues throughout by being generic.  Its long been true users could have just used the Row object and called it a day as that was the design.